refactor: move Cooldown class from context to candidate module
Cooldown is a data class closely related to candidate filtering, not
to the WorkContext lifecycle.  Moving it to fromager.candidate removes
a circular import workaround in resolver.py and keeps the candidate
module self-contained.
